Remove deployment files from Metalware repo
These are content rather than part of the core functionality of
Metalware, and so have been pulled out to repos to be used with
Metalware.

The main repo containing most of these files is
https://github.com/alces-software/metalware-profile-hpc; while
https://github.com/alces-software/metalware-default is a basic, minimal
repo. Further, some of these files related to setting up the current
machine as a Metalware deployment server have also since been extracted
to the Controller appliance in Imageware.