Implement custom argument validating
For some reason Commander doesn't include any validation that the
correct number of arguments are passed to a command.

This commit makes our commands use a custom `Command` class which:

a. validates the command `syntax` is in a consistent form of
   `metal COMMAND [ARGS...] [options]`;

b. validates the number of arguments passed to a command is at least the
number of required arguments, and no more than the total possible number
of arguments (required plus optional); if this is not the case an error
is shown.