Add tests for updated `hosts` command
These just test that the `hosts` command interacts with the `Templater`
and `Iterator` correctly, with the former mocked out, rather than going
the full way to test the templating itself. This has the following
advantages:

- We don't need to test the templating again for every command, when it
should just operate the same for all of them, we can just ensure we
invoke the correct `Templater` methods. So long as the `Templater`
itself works correctly everything else should work fine.

- If the tests for every command all used the real `Templater` we would
need to either mock real files or the filesystem interactions all over
the place; it is simpler to just mock the `Templater` itself everywhere
except when testing it directly.

Mocking the interactions with `nodeattr` also has the advantage of
allowing the current tests to all be run without `nodeattr` or `genders`
being set up.