Allow use of `alces` values within configs
This is required to allow changing the config based on the
current/passed/generated `alces` values, and is consistent with how
things worked with the previous global magic parameters.

A lot of things needed to be changed to allow this due to how our
templating works, including turning the `alces` object into a hash when
templating, which is a temporary change required as we convert the
current config to/from a hash using `to_s` and `eval` while parsing the
configs.