Raise error when cannot find gender group
In most cases this is what we want, as normally cannot operate with a
non-existent gender group.

When templating still return an empty array when attempt to access a
non-existent gender in the magic namespace, however this may change
later and/or we will give a warning in this case.