Prevent `build` `ensure` from suppressing errors
Previously the shared code to clean up after a successful build or an
interrupt was included within an `ensure` block, however if another
error occurs this code will still run, and can potentially raise a
different error itself. For instance this was happening if the config
fails to be created due to the config file not existing.

The fix is to use a shared method only where needed rather than relying
on `ensure`.