Refactor to use `Templater` via class methods
This makes using, as well as testing the use of, the `Templater`
simpler, as rather than needing to create `Templater` instances all over
the place and then invoke methods on these, we can just run a class
method with the correct arguments that takes care of this for us. In all
tests except those for the `Templater` itself we can then just assert
that the correct class method is invoked with the correct arguments.

This makes testing templating simpler in each case than what we
currently need to do, where we have to assert that a `Templater` is
created correctly, and then assert the correct created instance has the
correct instance method called on it.