Add method to `Node` to get `build` `files`
This gives the merged namespaces with arrays of file identifiers from
all `files` namespaces within all configs for the node, with identifiers
in higher precedence configs replacing those with the same basename in
lower precedence configs.

Also added the `Node#config` method which is used by `Node#build_files`;
this is public as it is likely it will be useful/used elsewhere soon.