Add initial rendering of local build files
Currently just attempt to render the local build files for a node (both
within the repo and absolute paths elsewhere on the Controller
appliance). Still need to handle remote files, specified by a URL, as
well as invalid local files and passing through the rendered files info
when rendering the other templates.