Avoid having multiple configs for unit tests
This avoids us having one Metalware config for use in most unit tests,
and another slightly different one for use in those tests where we want
to use our repo configs from the fixtures.

Instead we can make the path to the configs an independently
configurable parameter from the path to the repo; this way we can have
the fixtures configs always be used, while still always testing the
paths which would really be used (the `/var/lib/metalware/repo` paths)
are used.