Implement happy path for class to retrieve files
`BuildFilesRetriever#retrieve` takes a files hash, as given by
`Node#build_files`, and returns the hash of local paths and info on the
given build files.

This currently works, as long as every template path is always valid and
we never get an error when retrieving any URL. Since we want to just
warn rather than blow up in these cases error handling is still to come.