Pass `files` when rendering build templates
When rendering templates in `build` - for Kickstart, PXELINUX, and
arbitrary other files specified by the `files` config parameter - we
want to pass in the information on the retrieved `files` so this can be
referred to within the templates.

For example this will allow the Kickstart template to retrieve rendered
`files` to perform further customization, or the rendered `files`
themselves to dynamically refer to other rendered scripts or configs.