Remove invalid URL-specified file in test
In order to accurately distinguish file paths from URLs when specifying
build `files` in repo configs, we require the latter be specified with a
protocol (e.g. `http`). Therefore, this URL will not be interpreted as
such when retrieving the build `files` to use, but rather as a path
relative to the repo's `files` directory, and so including it is
misleading.