Adapt determining command name for subcommands
In `BaseCommand#command_name` we determine the unique name of the
current command from the class name for that command. Previously this
did not work for commands with subcommands as we only considered the
last part of the class name; this commit changes this so we include all
parts of the full name after the `Metalware::Commands` namespace,
meaning this now works for subcommands.