Test status monitor in rspec
Created a status spec test. The previous tests can not be ported as
they use a different testing module. They will need to be replicated
in rspec.

Added rspec tests for the status monitor. Also added a `spec/slow`
directory for test that take awhile (~1 sec) to run. They are not ran
with the unit-tests but instead with the integration tests.