Always read repo path from config
Previously the value for the path to the repo could come from two places
in different parts of the codebase: both from `Constants::REPO_PATH` and
`Config.repo_path`. Since we want to externally change this when testing
the latter is the better choice for this since this will allow us to do
this, therefore this commit changes all usage of the repo path to get
this from the config.

This does have the downside that we now need to pass the config through
to all templater render methods as the templater needs access to the
repo path, however I can't see a way around this other than making the
config global in some way, which I think would lead to more complexity.
This could later be made clearer however, for instance by using keyword
arguments or currying when passing arguments to the render methods.