Give descriptive error for unset template params
When an attempt is made to access a method of an unset template
parameter, this is an error as there is no way to handle this and
continue (without assuming some value for the unset template parameter,
which is probably not what we want to do).

In this situation we want to display an informative error making clear
that the issue lies with the template rather than Metalware itself; this
commit attempts to do this in a simple way.