Patch Commander to prevent global implicit short-tags
Commander will implicitly match short tags for global options. E.g.
--strict will cause metalware to operate in strict mode. However -s
will also do the same. The short tags are also not listed in the help
and are causing obsecure bugs as they are poluting the name-space.

This commit handles the first part of the patch. That is, Commander
removing the short tag from the CLI input preventing it from being
passed to the underling command. It also removes the -t option from
--trace CLI help as it clashes with -t TEMPLATE for other commands.

Note this does not stop Commander triggering the global option. E.G.
-t will be interpreted as both a template input and trace.