Render and run Each Command and create tests
Updated the Each CLI to take the group long tag. Updated Each setup
method to use Nodes which can be iterated over. Created a Templater
instance method that renders from a string instead of a file. Using
this new templater method, Each then renders the command and runs it
as a system call.

A rspec test suite has been created for Each as well. During testing,
stdout needs to be redirected which could only be acheived using a
temporary file. This is also means the Each Command supports file
redirection explicitly from ruby.