Allow iterating through template parameters
When writing Metalware templates it is sometimes desirable to be able to
iterate through a set of key-value pairs, rather than needing to access
parameters individually; previously this was not possible (without
writing a convoluted template).

This commit allows this by adding a new class to be used for the
template parameters loaded from the repo config, which functions the
same as `RecursiveOpenStruct` but with an additional `each` method.

This seemed the simplest way to add this functionality without making
future issues and confusion likely; we do not want to make this class
have all `Enumerable` methods otherwise issues would occur when setting
any properties with the same names in the config.