Move global options into yaml and add parser
Move the global options from cli.rb to config.yaml and added a parser
to parser.rb.

Fixe white-space formatting issues in the templates by removing
trailing endline characters added from the yaml parser.

Removed error if the calling object isn't recognised. Instead return
the yaml hash. This is required by for the autocomplete

Reduced the number of uses of `send` to only when required. When the
method is known in advance (via the logic), then it can be called
directly.