Pull autocomplete commands from yaml file
Autcomplete loads and traverses the cli yaml file instead of reading
the file structure. Intialization has been adjusted so that metal and
cur command do not appear as completed commands.

Autocomplition will be from now on based on the current_hash. This
hash corresponds to the list filled command. Additional logic is
included to prevent commands filling subcommands even if an option
(or junk) has been included on cli