Fix broken tests which load repo_configs_path
As a result of using load_config in Node class instead of building
of the repo bass path, a bunch of tests broke. This fixes them by
spoofing repo_configs_path.

Also ensure that test that are intended to work without a repo use
a empty repo dir. Previously they used the default repo location
which may or may not have files in it.