Integrate the Node answers into the MagicNamespace
The Node object is now directly passed into the MagicNamespace struct.
This means the nodename is nolonger passed in. Instead the nodename
is pulled from the Node object. The Node object is not passed into the
underling Struct object to prevent the users from trying to directly
access it.

The answers can be accessed with `alces.answers` which returns the
Node answers. The MissingParameterWrapper is also used for the answers
to detect when answer is missing.

The MissingParameterWrapper now accepts Hashes as a valid input which
will automatically be converted to a IterableRecursiveOpenStruct