Convert UnsetConfigLogError into a MetalwareError
Originally UnsetConfigLogError was directly inherited from Exception.
However this causes it to be missed by Commander and the full stack
trace displayed each time it is thrown. It is also typically the error
thrown if their is an issue with the config file.

This commit makes it a Metalware error which can be caught by Commander
and handled with the standard --trace flag