Integrate default answers to topology questions
Configurator reads the default answer to questions and optionally includes
it into the Question object (if it exists). Then when the questions are asked,
the default question is added depending on what type of default it is.

As Metalware needs to support erb tags in the default answer, Highline needed
to be patched. Highline previously did not sanitize the default value string
and thus tried to render the erb tags causing an error. The highline patch
escapes out erb tags in the default value preventing highline from erroring.

configurator_spec has also been altered slightly so that the test HighLine
reads from a temporary file instead of STDIN. This allows the testing of the
patch behaviour instead of stubbing out how highLine functioned.