Replace CLI HighLine with a new one in Configurator
Previously `self` was passed into Configurator, which was then used
as the highline object. However as HighLine was not require by Metalware
it is vague where the Highline object came from (likely Commander).

As HighLine needs to be altered so that it supports inputs with ERB tags,
(commits to come) a fresh copy of HighLine is created by default. The
tests require a highline input to function properly and therefore the
input has not been removed completely