Enforce Repo dependencies in the Dependencies class
Created the Dependencies class that will for now enforce dependencies
between the commands. This commit implements the repo dependencies,
making the Repo class obsolute.

Dependencies for a Command are set by overriding the dependencies_hash
method of the BaseCommand. It needs to return a hash which encodes the
dependencies to enforce. At this point in time it only accepts a repo
key to enforce repo dependencies. A value of true means the repo must
exist, alternatively a array of strings indicate which directories must
exist in the repo. Dependencies does not check if the template files
exist.

Minor changes had to be made to existing test to mock out the new
Dependencies class. A load_repo method has been added to FakeFSHelper
that takes path to a directory. This directory is then loaded in the
repo_path location as specified by metal_config. If no path is included
then the repo as specified by config.repo_path is loaded.