Pendantry changes as discussed in #90
Simplified the CHANGELOG.

Moved the test <answer>.yaml files into their
correct locations (previosly their was no groups or nodes dirs). Updated
Node to read the answer files from their correct locations.

Made @fshelper a local varaible in the tests where possible.

Stubbed out Node object in the Templater tests to reduce coupling.

Revert the passed_magic_parameters back to its previous select method
and then tap on the changes. This should simplify the code.

No longer pass the Node object into the MagicNamespace struct. It is
only intialized with it and then saves it as a private instance method.

Their is the intial code to cause fatal erros in the MissingParameterWrapper
but this has been commented out as it is causing tests to fail otherwise