Throw a fatal error if a configure answer is unset
The MissingParameterWrapper now has the option of throwing an error
instead of a warning. The intial code to complete this was included
in the previous commit, but wasn't fully implemented as it was causing
errors in the Templater tests.

Due how the Templater was mocked, it was accessing parameter it should
have been and thus causes an error. This commit updates the templater
tests to use FakeFS. A new clone_repo command has been included in
FakeFSHelper to assist creating the repo in the fake FS