Generalize Dependencies to check files existance
Instead of checking the directory structure of repo, Dependencies
goes throw two stages: 1) Check if the base repo exists, 2) check
if the file path exists inside the repo. During step 1 it checks
that the repo is a directory and step 2 if the relative repo path
is a file (not a directory).

This 2 stage structure will then be used to validate configure paths