Enforce configure dependencies on Build
Implemented Configure Dependencies which is responsible for checking
that the answer files have been generated. The first stage is to
validate configure in general, which checks the repo and answer
directories exist.

The second stage is to check that the file paths passed into with the
Configure Dependencies option exist. Note, inorder to maintain backwards
compatability, this is skipped if the `configure.yaml` file is missing
from the repo.

Dependencies code has also been simplified by merging the path_exists?
and generate_file_path functions into file_exists?. Note that if `true`
is included as a path, it will perform the first stage check (validate
configure/ repo) but skip the file checking.

All Confugure commands require the answers directory to exist. It also
requires configure.yaml to exist in the repo as its dependencies.

The Build command depends on `domain.yaml` existing as an answer as well
as groups/<group>.yaml if -g is included. NOTE: the files dependency
will only be inforced if `configure.yaml` exists in the repo.