Record primary group for each new group configured
This is the first step towards having a consistent `group_index` for
each new group configured using `metal configure group`.

Every group created in this way is now added to a `primary_groups` list
in a new `groups.yaml` cache file. When it is later added, the
`group_index` for a given primary group will then be the index of the
given group within this array; it will therefore be guaranteed to be
consistent as new groups are added (barring this file being modified
outside of Metalware).