Fix cli integers being converted to octal
By default, Integer inputs into `Commander` will be interpreted as
different types (e.g. decimal, hex, octal) depending on their prefix.
This commit forces all `Integer`s to be interpreted as Decimal. Other
input types can still be used, but must be specified explicitly.

It also uses version 4.5.2 of `Commander`. This version implements
a patch of the Decimal type. Previously even if a number was explicitly
stated as Decimal, it could still be interpreted as Octal. This
update prevents that.