Update templater tests to use FileSystem class
As well as replacing the use of `FakeFSHelper` in the `Templater` tests,
this also changes some tests where we stub the use of a new config file.
Instead, these now use the standard config and the fake filesystem,
which should better test that things are actually working and make
things clearer.

To have this work well I also needed to add the ability to setup the
`FileSystem`, without letting it get into an inconsistent state or
having the `FakeFS` be used prior to running `test`; see the comment in
`FileSystem` for details on this.