Standardise dependency input value to be an array
Previously the value associated with a dependency could be a string,
array or True which required additional logic to handle it. From now
on, the value must be an array. An empty array can be used to check
the dependency group without individual values. Thus the empty array
replaces True.