Remove magic returned variable in dependencies
Previously repo_configure could return a value "skip" which indicated
note to enforce the dependency file paths. However this added unecessary
complexity. This has now been removed and the `Dependency` code
simplified.

As a result, `Dependencies` is no longer fully backwards compatible
with repos that do not include topology questions. This also broke
the `build` integration tests. To fixes them the cloned repo needed
to be updated. Please note, this is only a temporary fix, the exact
branch that is cloned should be made into a tag so that it is static.

Also as `Dependencies` checks if answer files exist, the answer file
path needed to be changed in the integration tests. This means that
answer_files_path is now a config parameter instead of a Constant.