Update Build integration test to use static answers
Now that build dependencies use static answers and use the new
templates, an error occurs if cluster_name has not been set. This
commit creates static answer files to be used by the tests.

It also updates the Gemfile.lock to use version 4.5.3 which fixes
a bug in the OptionParser patch.

Note: that the build integration tests are failing. This is because
they are currently failing in develop at the time this branch was
rebased.