Fix bug where bash commands strip quotes
As both ipmi and power are bash scripts that are ran using BashCommand
and bash, the quotes need to survive both stages. Previously quoted
will be stripped cli and become seperate arguments in ARGV. However
if their are spaces, ruby will pass them to bash as seperate arguments.

This commit fixes the bug by directly passing the input arguments as
an array.