Enforce required and allow optional questions
Configure questions are now required by default. This means that the
user will be prompted to renter an answer if they do not give one and
it does not have a default.

A new optional tag can be included with the questions, which means
an empty input is a valid response. The optional tag must evaluated
to a truthy value.

Note that when HighLine re-prompts for an input, it uses inspect to
generate the message. This gives crazy responses such as:
`must match <proc:/some/file/path-352random-other-stuff>`

As this isn't very useful to the user, a HighLinePrettyValidateProc
class is used. This class overrides the inspect method so that it
gives nicer prompts to the user.