Introduce minimal repo for integration tests
Our integration tests were failing due to the use of the
`metalware-default` repo.

Firstly, this was only being cloned if is was not already present (to
prevent needing to clone it each time these tests were run). This meant
the repo was not being updated and the code to do this had several
issues: an old local version of the repo was being used, the code to
clone it had broken (it referred to `SystemCommand` rather than
`Metalware::SystemCommand`), and it also referred to a
no-longer-existing branch.

Additionally, once the latest `master` of this repo was used tests would
then fail due to the `files` specified in this repo referring to various
URLs, which would cause these to all be attempted to be retrieved,
meaning commands took too long to run and would not complete the
required behaviour being tested.

The solution is to introduce a minimal repo containing only the basic
required repo structure, to be used in these and any future integration
tests. Since this now lives in this repo it should always be up-to-date
with the code using it, and so hopefully prevent similar future issues.