Add `configure.yaml` to fix tests after rebase
A minimal `configure.yaml` is needed in the integration test repo, to
fix the tests after the rebase on `develop`. `build` now has a
dependency on this file existing, and `develop` no longer tests using
the `metalware-default` repo so this file did not exist in the repo we
now test with.