Make testing build files not require unit test config
Hiding the stubbing of a constant to use the unit test config, when the
method to create a build files hash for testing is called, is confusing
and prevents this method from being used easily when we want to use
`FakeFS` rather than faking file paths. This commit removes this and has
requires the config to use be passed in instead.