Implement `Node#group_index`
This returns the index of a node's primary group, which is cached when
the group is configured and so should remain consistent for all commands
unless the cache file is manually edited. If the node's primary group
does not appear in the cache then an error is raised, as the primary
group for each node should be configured prior to any attempt to use it.