Validate configure.yaml with dry::validation
This commit contains the base code used for the dry validation. Note
this commit has been rebased on an earlier attempt to implement
validation which was ultimately unsucessfull. The tests however have
been salvaged.

This commit also adds some copyright notices and mocks out validation
in the other tests (as it currently will break them).

Please ignore the code in Validator::Configure class at this stage.
It is mostly going to change but provides the structure for future
commits.