Validate String questions in Configure.yaml
Adds the logic to validate the String questions in configure.yaml.
Currently only string answers are supported by the validator. The
configure.yaml file also requires the domain, group and node feilds.
An additional questions field is also allowed, but no validation is
performed on it. All other fields will cause an error.

The other-configure files have been removed as the commands are now
being testsed directly with hashes for ease of maintaince.