Have all YAML interactions use Data module
All handling of our YAML data files is now done through a standard
module, simplifying this and meaning there is only one place to change
how we do this.

The one exception is the `CliHelper::Parser`, which is intended to load
a CLI specification file edited by developers only, rather than interact
with Metalware data files, so it seemed best not to couple this to how
Metalware itself does its data file handling.