Create new templating directory
Move classes only used when templating into a new directory for these,
to keep the top-level `src` directory a bit clearer. `Templater` itself
is not moved here as it is used all over the place and we want to be
able to refer to it as `Templater` rather than namespaced under a new
module name (which will be introduced for this directory in a future
commit rather than this so Git picks up that these files have moved
rather than been deleted).