Setup two stage validation of configure.yaml
Due to the identifiers being user generated, they can not be used
by Dry::Validation as they are not known in advance. The solution
to this is to split the validation in two parts.

The first stage is to validate the base structure of the file, making
sure it has the three section: domain, group and nodes. The second
stage is to loop through each question all the sections and validate
the question is correct.

If an error occurs at any point, the particular validation results
containing the error is returned. If however they all succeed, the
configure (stage 1) results are returned to show no error has occured.

The stage 1 validation has been completed, and the base code for the
stage 2 has been setup. Also move the error messages to an 'errors.yaml'
file.