Validate the base structure of the questions
Loops throught the questions creating a new hash with the info
contained within it. The reason the section and identifier are
included but not validate is so they appear along with any error
messages. They have however allready been validated and thus do
not need to be again.

The fields in the question are white-listed and any required fields
are now enforced. The basic return type for each field is also checked.
This commit, also introduces the idea of the supported type. These
are the type of answers that can be supplied to the question.

Please note that 'choices' has not been validated as it is significantly
more tricky and will be addressed latter.