Implement type checking of default values
This commit introduces type checking for default values. Currently
only Integers and Strings support defaults. If an default is included
for a Boolean, it will throw an error.

The type checking needs to be done by using `rules` as they are
dependant on the value of the `type` fields. Rules work on linear
algebra but using the predicates as defined by Dry::Validation.