Introduce PrimaryGroup class
This is for interacting with the cached configured primary groups.
Currently a `PrimaryGroup` instance is a simple object with just a name
attribute, and the class provides Enumerable methods for instantiating
these from the cached primary groups.

There is some overlap now between the `PrimaryGroup`, `GroupNamespace`,
and `Nodes` classes, and there is probably a better way the behaviour of
these could be divided up; I am not quite sure what the best way to do
this is at the moment so leaving this for now.