Make node template parameters available for group
Can now iterate through each node in a group via the `group.nodes`
method, where `group` is a templating namespace for a primary group as
yielded to the `alces.groups` method. For each node the full templating
config is available, as would be made available when rendering a
template for just that node.