Convert boolean strings to booleans when templating
This makes templating using boolean values defined in terms of other
booleans work a bit more as expected.

Without this, if ERB was used to specify a boolean value in the config
in terms of another value, then a string would be used instead. This is
consistent with how the templating always works but would lead to
silently unexpected results, as "false" is truthy.

Special-casing this fixes this, although it is a bit ugly. We may want
to perform similar conversion of other values, like integers. Or
possibly not, as there is no way of knowing at this point if the string
"5" should be interpreted as "5" or 5.

In defense of just doing this for booleans, using most other values
incorrectly should normally lead to obvious errors, e.g. `<%= "5" + 5
%>` will raise an exception; this can then be noticed and the desired
conversion can easily be done using `to_i` or `to_s`.