Replace stubbing out HighLine and use actual input
Rather than faking various HighLine methods, now we have a setup to use
actual input for the asked questions we can just do this instead. This
simplifies things and should be a more thorough test.

We still check the correct HighLine methods are called in a couple of
places where how this is done matters, but we can then call the original
method to ensure this works as expected as well.