Allow defaults and saved answers for boolean questions
These weren't done before as this is slightly more tricky to configure
than defaults for other types of questions.

As a side effect of doing this readline bindings now work for boolean
questions as well, as they now yield to the block setting these up in
the same way as other questions.