Make ConfigureCommand test slightly more robust
Override the highest level method which needs to be overridden for the
test to work (this needs to be done as this test class is not in the
right namespace); this way the lower level methods which depend on this
will still run and be tested.

The test class name can now also be changed back, making its purpose
more obvious.