Include present answers always
We should still include the answers when not templating for a node,
although we don't want to raise an error if one is not present, as
otherwise we won't be able to use any answers when e.g. rendering domain
templates.

This is part of resolving https://trello.com/c/TsFiI7rL.