Handle passing validation block to render_to_file
If a block is given, pass it the rendered template and it should return
whether this is valid and should be written to the file.

This is intended for use when rendering the genders file, so we can
validate it before installing it; it may also be useful for rendering
the dhcp template or others.