Have genders be rendered to/used from data dir
Having the `genders` file be rendered to and used from the location
Metalware is installed to causes issues in development, as the
development Metalware renders its own genders within `/tmp/`, but
`nodeattr` is setup to use the `genders` within `/opt/`. Having the
`genders` file in a consistent location prevents this issue, and is also
consistent with everything else we render.