Implement view-answers commands
These are mostly simple wrappers around `AnswersTableCreator` functions;
each displays a table showing the answers to the `configure` questions
for the given object, along with any other applicable answers which that
object will fall back to using if its own answer is unset.

Aside from this, each command sets the `dependency_hash` to ensure the
prerequisites for its answers to be set exist.

`ViewConfig::Node` also checks the node is a valid configured node, i.e.
it is part of a configured group (each node managed by Metalware must be
part of a configured primary group; the node itself does not need to be
configured, this just allows further customisation).