Allow building non-primary groups
This code in the build dependencies was meaning we required that the
group to be built has an answers file, i.e. is a configured primary
group. However, this prevents building non-primary groups as they will
never have an answers file. Just removing this should be fine; I can't
see any issue which could occur by not having this.