Implement integer type checking of answer validator
This commit contains the base code for type checking of the answer
files. Because of the complexity of the validation, it has been
broken down into multiple steps.

The first two steps validate the structure of the answer file. In
the process, a validation_hash is created in a format that is easier
to validate with Dry::Validation.

The lasy step is looping through all the questions and checking the
type. At this point in time only integers can be validated. All
other types will assumed to be correct (this will be changed to
assumed to be incorrect, see TODO).

This commit does not contain the code to integrate in with Dependency
which will be completed at a latter date. It also contains an update
to the Gemfile.lock, .gitignore and Makefile which was required to get
rubocop working locally on my machine.