Remove hosts command from Metalware
With the addition of the `hosts` file being rendered in full as
`configure` commands are run, the `hosts` command is now obsolete, and
in fact always errored currently due to the changed template. This
commit removes it, its tests, and the
`Templater.render_and_append_to_file` method only used by it.