Stop saving of default configure question answers
When configuring, if a user enters no answer for a question which has a
default we want to proceed with the configuration process (as we did
already) but we do not want to save the default as an answer.

This means that the default answer for a question as set in
`configure.yaml` will always continue to be used, so long as a user has
not explicitly set their own answer, whatever the default might later be
changed to (at least, this is what should happen once we have included
the defaults when forming the overall answers for a domain/group/node).