Bump integration test timeout
An attempt to try to fix intermittent failure of 'handles "no" to
interrupt prompt' test. `git bisect` tells me that
510d152b291037bce21fe57224d98151cdf6746d is the commit which broke this,
but I can't tell why - from manually checking the addition of 'rubocop'
to the Gemfile appears to make the difference between whether this
consistently or intermittently fails. Bumping this timeout appears to
make it pass consistently again, for me at least.