Dynamically decide namespace fields to hash
This commit introduces a module with a single function which will take
an object and convert its methods and their values to the keys and
values of a hash, with the optional ability to override what method is
used to get certain fields. This function is then used in the
`MagicNamespace` and `GroupNamespace` when we convert them to hashes, so
we don't need to manually specify the fields to include.

This function will of course only work for objects where all methods to
be included can take no arguments, otherwise it will blow up as the
methods won't be sent enough arguments.