Allow optional files in Dependency
A new optional hash can be included in the dependency_hash which
indicates files that should be validated but does not throw an error
if they do not exist.

The optional hash has the same syntax as a regular dependency_hash.
At this point in time, the superfluous, however it will be needed
once Answer Validation is included.

Also adjusted Makefile and .gitignore so that I can run rubocop
locally