Include defaults when forming answers
Previously we didn't need to do this as we were always saving question
defaults, so files in the answers hierarchy would always have these
values set. However, now we only save answers different to the default
or higher level answer when configuring, meaning we need to include the
defaults when forming the answers to be used so these values will be
present in the final answers hash.

Note: the deleted file is deleted because this file is an invalid
`configure.yaml` file, and now that we are loading `configure.yaml` when
forming the answers its existence was causing various tests in other
places to fail; by removing this the minimal valid `configure.yaml` file
from `MinimalRepo` is used wherever this was used instead, meaning these
tests pass again.