Implement nodes_in_primary_group command
Previously it was only possible to get a complete list of nodes in
a particular gender group. However this meant it isn't possible to
get a list of nodes that are in a particular primary group.

This commit introduces nodes_in_primary_group which assumes the first
group listed for a node in the genders file is the primary group.