Validate when loading/dumping data files
The YAML data files Metalware works with should always contain hashes,
so this commit checks the data we try to load/dump is valid.

Previously these methods would still error in this case but with an
uninformative error due to `deep_transform_keys` not being a valid
method on the object; I thought giving an informative error is better
than this, and also better than changing to be more lenient with the
data we load/dump, since we always want this to be hashes anyway.