Remove validating answers file contains a hash
We now always validate when loading and dumping data using the `Data`
module that the data to be loaded/dumped is a hash, as this is what we
always want to work with in Metalware. Therefore, this answers
validation is unnecessary as by the time it is run the data it is
validating is guaranteed to be a hash; if the data is not a hash
Metalware will already have exited and displayed an informative error to
the user.